fix(routing): move Super Admin pages under /cipp/advanced/ for consistent URL paths
All Advanced menu children use /cipp/advanced/... except Super Admin,
which used /cipp/super-admin/.... This moves the pages and updates all
route references, relative imports, and navigation configs to align
Super Admin with its sibling menu items.
