+++ src/layouts/config.js
@@ -1012,7 +1012,7 @@ export const nativeMenuItems = [
         items: [
           {
             title: 'Super Admin',
-            path: '/cipp/super-admin/tenant-mode',
+            path: '/cipp/advanced/super-admin/tenant-mode',
             roles: ['superadmin'],
             permissions: ['CIPP.SuperAdmin.*'],
             scope: 'global',
