+++ src/pages/cipp/advanced/super-admin/cipp-roles/edit.js
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
-import { Layout as DashboardLayout } from "../../../../layouts/index.js";
-import CippPageCard from "../../../../components/CippCards/CippPageCard";
-import { CippRoleAddEdit } from "../../../../components/CippSettings/CippRoleAddEdit";
+import { Layout as DashboardLayout } from "../../../../../layouts/index.js";
+import CippPageCard from "../../../../../components/CippCards/CippPageCard";
+import { CippRoleAddEdit } from "../../../../../components/CippSettings/CippRoleAddEdit";
 import { CardContent, Stack, Alert } from "@mui/material";
 
 const EditRolePage = () => {
