+++ src/pages/cipp/advanced/super-admin/sam-app-roles.js
@@ -1,15 +1,15 @@
-import { TabbedLayout } from "../../../layouts/TabbedLayout";
-import { Layout as DashboardLayout } from "../../../layouts/index.js";
+import { TabbedLayout } from "../../../../layouts/TabbedLayout";
+import { Layout as DashboardLayout } from "../../../../layouts/index.js";
 import tabOptions from "./tabOptions";
-import CippFormPage from "../../../components/CippFormPages/CippFormPage";
+import CippFormPage from "../../../../components/CippFormPages/CippFormPage";
 import { Alert, CardContent, Stack, Typography } from "@mui/material";
 import { WarningAmberOutlined } from "@mui/icons-material";
 import { useForm } from "react-hook-form";
-import { ApiGetCall, ApiGetCallWithPagination } from "../../../api/ApiCall";
+import { ApiGetCall, ApiGetCallWithPagination } from "../../../../api/ApiCall";
 import { useEffect } from "react";
-import CippFormComponent from "../../../components/CippComponents/CippFormComponent";
-import GDAPRoles from "../../../data/GDAPRoles";
-import { CippFormTenantSelector } from "../../../components/CippComponents/CippFormTenantSelector";
+import CippFormComponent from "../../../../components/CippComponents/CippFormComponent";
+import GDAPRoles from "../../../../data/GDAPRoles";
+import { CippFormTenantSelector } from "../../../../components/CippComponents/CippFormTenantSelector";
 
 const Page = () => {
   const pageTitle = "SAM App Roles";
