+++ src/pages/cipp/advanced/super-admin/tabOptions.json
@@ -1,0 +1,26 @@
+[
+  {
+    "label": "Tenant Mode",
+    "path": "/cipp/advanced/super-admin/tenant-mode"
+  },
+  {
+    "label": "Function Offloading",
+    "path": "/cipp/advanced/super-admin/function-offloading"
+  },
+  {
+    "label": "Time Settings",
+    "path": "/cipp/advanced/super-admin/time-settings"
+  },
+  {
+    "label": "CIPP Roles",
+    "path": "/cipp/advanced/super-admin/cipp-roles"
+  },
+  {
+    "label": "SAM App Roles",
+    "path": "/cipp/advanced/super-admin/sam-app-roles"
+  },
+  {
+    "label": "SAM App Permissions",
+    "path": "/cipp/advanced/super-admin/sam-app-permissions"
+  }
+]
