+++ src/pages/cipp/advanced/super-admin/time-settings.js
@@ -1,12 +1,12 @@
-import { TabbedLayout } from '../../../layouts/TabbedLayout'
-import { Layout as DashboardLayout } from '../../../layouts/index.js'
+import { TabbedLayout } from '../../../../layouts/TabbedLayout'
+import { Layout as DashboardLayout } from '../../../../layouts/index.js'
 import { useForm } from 'react-hook-form'
 import { Grid } from '@mui/system'
-import CippFormComponent from '../../../components/CippComponents/CippFormComponent'
-import { ApiGetCall } from '../../../api/ApiCall'
+import CippFormComponent from '../../../../components/CippComponents/CippFormComponent'
+import { ApiGetCall } from '../../../../api/ApiCall'
 import { useEffect } from 'react'
-import CippFormPage from '../../../components/CippFormPages/CippFormPage'
-import { useTimezones } from '../../../hooks/use-timezones'
+import CippFormPage from '../../../../components/CippFormPages/CippFormPage'
+import { useTimezones } from '../../../../hooks/use-timezones'
 import tabOptions from './tabOptions'
 import { Alert, Typography } from '@mui/material'
