--- src/pages/cipp/super-admin/tabOptions.json
@@ -1,26 +1,0 @@
-[
-  {
-    "label": "Tenant Mode",
-    "path": "/cipp/super-admin/tenant-mode"
-  },
-  {
-    "label": "Function Offloading",
-    "path": "/cipp/super-admin/function-offloading"
-  },
-  {
-    "label": "Time Settings",
-    "path": "/cipp/super-admin/time-settings"
-  },
-  {
-    "label": "CIPP Roles",
-    "path": "/cipp/super-admin/cipp-roles"
-  },
-  {
-    "label": "SAM App Roles",
-    "path": "/cipp/super-admin/sam-app-roles"
-  },
-  {
-    "label": "SAM App Permissions",
-    "path": "/cipp/super-admin/sam-app-permissions"
-  }
-]
